move around UART config + io threads
all this code can live under the UART device itself.
