+++ src/devices/uart/mod.rs
@@ -1,0 +1,7 @@
+mod cfg;
+mod device;
+
+pub use cfg::*;
+pub use device::*;
+
+mod iothreads;
